Fix: 그룹원 초대 모달 WARNING 해결
#40
